refactor: 알림 컨테이너 로직을 커스텀 훅으로 분리 (#264)
