+++ frontend/src/components/layout/SidebarAlarm/SidebarAlarm.tsx
@@ -1,58 +1,23 @@
-import { useEffect, useRef, useState } from 'react'
+import { useRef, useEffect } from 'react'
 import clsx from 'clsx'
 import { Icon } from '../../ui/Icon'
 import { AlarmCard } from './AlarmCard'
-import { useSidebarAlarmStore } from '@/stores/useSidebarAlarmStore'
-import { useRequestStore } from '@/stores/useRequestStore'
-import { getRelativeDate } from '@/utils/getRelativeDate'
-import { ENDPOINTS } from '@/constants/url'
 import { AlarmMessage, AlarmType } from '@/constants/AlarmMessage'
-
-type Alarm = {
-  id: number
-  title: string
-  content: string
-  isChecked: boolean
-  createdAt: string | null
-  targetType: string
-  targetId: number
-}
-
-type AlarmGroup = {
-  type: 'all' | 'bookmark'
-  alarms: Alarm[]
-}
+import { getRelativeDate } from '@/utils/getRelativeDate'
+import { useSidebarAlarmLogic } from './useAlarmLogic'
 
 export const SidebarAlarm = () => {
-  const isOpen = useSidebarAlarmStore((state) => state.isOpen)
-  const close = useSidebarAlarmStore((state) => state.close)
+  const {
+    isOpen,
+    close,
+    activeTab,
+    setActiveTab,
+    handleAlarmClick,
+    filteredAlarms,
+    isLoading,
+  } = useSidebarAlarmLogic()
+
   const SidebarAlarmRef = useRef<HTMLDivElement>(null)
-
-  const [activeTab, setActiveTab] = useState<AlarmType.ALL | AlarmType.BOOKMARK>(AlarmType.ALL)
-  const [alarmsData, setAlarmsData] = useState<AlarmGroup[]>([])
-
-  const { getData, patchData, isLoading } = useRequestStore()
-
-  useEffect(() => {
-    const fetchAlarms = async () => {
-      try {
-        const res = await getData<{ success: boolean; message: string; data: AlarmGroup[] }>(ENDPOINTS.ALARM)
-        if (res.success) {
-          setAlarmsData(res.data)
-        }
-      } catch (error) {
-        console.error('알림 가져오기 실패:', error)
-      }
-    }
-
-    if (isOpen) fetchAlarms()
-  }, [isOpen, getData])
-
-  useEffect(() => {
-    if (isOpen) {
-      setActiveTab(AlarmType.ALL)
-    }
-  }, [isOpen])
 
   useEffect(() => {
     const handleClickOutside = (event: MouseEvent) => {
@@ -63,30 +28,6 @@ export const SidebarAlarm = () => {
     document.addEventListener('mousedown', handleClickOutside)
     return () => document.removeEventListener('mousedown', handleClickOutside)
   }, [isOpen, close])
-
-  const handleAlarmClick = async (alarm: Alarm) => {
-    if (!alarm.isChecked) {
-      try {
-        await patchData(ENDPOINTS.ALARM_CHECK(alarm.id))
-        setAlarmsData((prev) =>
-          prev.map((group) =>
-            group.type === tabType
-              ? {
-                  ...group,
-                  alarms: group.alarms.map((a) => (a.id === alarm.id ? { ...a, isChecked: true } : a)),
-                }
-              : group,
-          ),
-        )
-      } catch (e) {
-        console.error('알림 확인 실패:', e)
-      }
-    }
-    close()
-  }
-
-  const tabType = activeTab === AlarmType.ALL ? 'all' : 'bookmark'
-  const filteredAlarms = alarmsData.find((group) => group.type === tabType)?.alarms ?? []
 
   const wrapperClass = clsx(
     'absolute top-0 right-0 h-full w-full bg-alarmBarBg z-50',
